Split ListView and ListViewLayout info separate files.
