+++ Listable/Sources/Internal/Layout/ListViewLayout.LayoutInfo.swift
@@ -1,406 +1,11 @@
 //
-//  ListViewLayout.swift
+//  ListViewLayout.LayoutInfo.swift
 //  Listable
 //
-//  Created by Kyle Van Essen on 9/23/19.
+//  Created by Kyle Van Essen on 11/19/19.
 //
 
-import UIKit
-
-
-protocol ListViewLayoutDelegate : AnyObject
-{
-    func listViewLayoutUpdatedItemPositions(_ collectionView : UICollectionView)
-    
-    func heightForItem(at indexPath : IndexPath, in collectionView : UICollectionView, width : CGFloat, layoutDirection : LayoutDirection) -> CGFloat
-    func layoutForItem(at indexPath : IndexPath, in collectionView : UICollectionView) -> ItemLayout
-    
-    func hasListHeader(in collectionView : UICollectionView) -> Bool
-    func heightForListHeader(in collectionView : UICollectionView, width : CGFloat, layoutDirection : LayoutDirection) -> CGFloat
-    func layoutForListHeader(in collectionView : UICollectionView) -> HeaderFooterLayout
-    
-    func hasListFooter(in collectionView : UICollectionView) -> Bool
-    func heightForListFooter(in collectionView : UICollectionView, width : CGFloat, layoutDirection : LayoutDirection) -> CGFloat
-    func layoutForListFooter(in collectionView : UICollectionView) -> HeaderFooterLayout
-    
-    func layoutFor(section sectionIndex : Int, in collectionView : UICollectionView) -> Section.Layout
-    
-    func hasHeader(in sectionIndex : Int, in collectionView : UICollectionView) -> Bool
-    func heightForHeader(in sectionIndex : Int, in collectionView : UICollectionView, width : CGFloat, layoutDirection : LayoutDirection) -> CGFloat
-    func layoutForHeader(in sectionIndex : Int, in collectionView : UICollectionView) -> HeaderFooterLayout
-    
-    func hasFooter(in sectionIndex : Int, in collectionView : UICollectionView) -> Bool
-    func heightForFooter(in sectionIndex : Int, in collectionView : UICollectionView, width : CGFloat, layoutDirection : LayoutDirection) -> CGFloat
-    func layoutForFooter(in sectionIndex : Int, in collectionView : UICollectionView) -> HeaderFooterLayout
-    
-    func columnLayout(for sectionIndex : Int, in collectionView : UICollectionView) -> Section.Columns
-}
-
-
-class ListViewLayout : UICollectionViewLayout
-{
-    //
-    // MARK: Properties
-    //
-    
-    unowned let delegate : ListViewLayoutDelegate
-    
-    var appearance : Appearance {
-        didSet {
-            guard oldValue != self.appearance else {
-                return
-            }
-                                    
-            self.invalidateEntireLayout()
-            
-            switch self.appearance.underflow {
-            case .alwaysBounceVertical(let bounce):
-                switch appearance.direction {
-                case .vertical:
-                    self.collectionView?.alwaysBounceVertical = bounce
-                    self.collectionView?.alwaysBounceHorizontal = false
-                case .horizontal:
-                    self.collectionView?.alwaysBounceVertical = false
-                    self.collectionView?.alwaysBounceHorizontal = bounce
-                }
-            case .pinTo(_):
-                fatalError("Other types of underflow are not yet implemented.")
-            }
-        }
-    }
-    
-    //
-    // MARK: Initialization
-    //
-    
-    init(
-        delegate : ListViewLayoutDelegate,
-        appearance : Appearance
-    )
-    {
-        self.delegate = delegate
-        self.appearance = appearance
-        
-        self.layoutResult = LayoutInfo()
-        self.previousLayoutResult = self.layoutResult
-        
-        self.changesDuringCurrentUpdate = UpdateItems(with: [])
-        
-        super.init()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError() }
-    
-    //
-    // MARK: Querying The Layout
-    //
-    
-    func positionForItem(at indexPath : IndexPath) -> ItemPosition
-    {
-        let item = self.layoutResult.item(at: indexPath)
-        
-        return item.position
-    }
-    
-    //
-    // MARK: Private Properties
-    //
-    
-    private var layoutResult : LayoutInfo
-    private var previousLayoutResult : LayoutInfo
-    
-    private var changesDuringCurrentUpdate : UpdateItems
-    
-    //
-    // MARK: Invalidation & Invalidation Contexts
-    //
-    
-    func invalidateEntireLayout()
-    {
-        self.neededLayoutType = .requeryDataSourceCounts
-    }
-    
-    private(set) var shouldAskForItemSizesDuringLayoutInvalidation : Bool = false
-    
-    func setShouldAskForItemSizesDuringLayoutInvalidation()
-    {
-        self.shouldAskForItemSizesDuringLayoutInvalidation = true
-    }
-    
-    override class var invalidationContextClass: AnyClass {
-        return InvalidationContext.self
-    }
-    
-    override func invalidateLayout()
-    {
-        super.invalidateLayout()
-        
-        if self.shouldAskForItemSizesDuringLayoutInvalidation {
-            self.invalidateEntireLayout()
-            self.shouldAskForItemSizesDuringLayoutInvalidation = false
-        }
-    }
-    
-    override func invalidateLayout(with context: UICollectionViewLayoutInvalidationContext)
-    {
-        let view = self.collectionView!
-        let context = context as! InvalidationContext
-        
-        super.invalidateLayout(with: context)
-        
-        // Handle Moved Items
-        
-        if
-            let from = context.previousIndexPathsForInteractivelyMovingItems,
-            let to = context.targetIndexPathsForInteractivelyMovingItems
-        {
-            let from = from[0]
-            let to = to[0]
-            
-            let item = self.layoutResult.item(at: from)
-            item.liveIndexPath = to
-                    
-            self.layoutResult.move(from: from, to: to)
-            
-            if from != to {
-                context.performedInteractiveMove = true
-                self.layoutResult.reindexLiveIndexPaths()
-            }
-        }
-        
-        // Handle View Width Changing
-        
-        context.widthChanged = self.layoutResult.shouldInvalidateLayoutFor(newCollectionViewSize: view.bounds.size)
-        
-        // Update Needed Layout Type
-                
-        self.neededLayoutType.merge(with: context)
-    }
-    
-    override func invalidationContext(forBoundsChange newBounds: CGRect) -> UICollectionViewLayoutInvalidationContext
-    {
-        let context = super.invalidationContext(forBoundsChange: newBounds) as! InvalidationContext
-        
-        context.updateHeaders = self.appearance.layout.stickySectionHeaders
-        
-        return context
-    }
-    
-    override func invalidationContextForEndingInteractiveMovementOfItems(
-        toFinalIndexPaths indexPaths: [IndexPath],
-        previousIndexPaths: [IndexPath],
-        movementCancelled: Bool
-    ) -> UICollectionViewLayoutInvalidationContext
-    {
-        precondition(movementCancelled == false, "Cancelling moves is currently not supported.")
-        
-        self.layoutResult.reindexLiveIndexPaths()
-        self.layoutResult.reindexDelegateProvidedIndexPaths()
-                                
-        return super.invalidationContextForEndingInteractiveMovementOfItems(
-            toFinalIndexPaths: indexPaths,
-            previousIndexPaths: previousIndexPaths,
-            movementCancelled: movementCancelled
-        )
-    }
-    
-    override func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool
-    {
-        return true
-    }
-    
-    private final class InvalidationContext : UICollectionViewLayoutInvalidationContext
-    {
-        var updateHeaders : Bool = false
-        var widthChanged : Bool = false
-        
-        var performedInteractiveMove : Bool = false
-    }
-    
-    //
-    // MARK: Preparing For Layouts
-    //
-    
-    private enum NeededLayoutType {
-        case none
-        case updateHeaders
-        case relayout
-        case requeryDataSourceCounts
-        
-        mutating func merge(with context : UICollectionViewLayoutInvalidationContext)
-        {
-            let context = context as! InvalidationContext
-            
-            let requeryDataSourceCounts = context.invalidateEverything || context.invalidateDataSourceCounts
-            let needsRelayout = context.widthChanged || context.performedInteractiveMove
-            
-            if requeryDataSourceCounts {
-                self = .requeryDataSourceCounts
-            } else if needsRelayout {
-                switch self {
-                case .none, .updateHeaders: self = .relayout
-                case .relayout, .requeryDataSourceCounts: break
-                }
-            } else if context.updateHeaders {
-                switch self {
-                case .none: self = .updateHeaders
-                case .updateHeaders, .relayout, .requeryDataSourceCounts: break
-                }
-            }
-        }
-        
-        mutating func update(with success : Bool)
-        {
-            if success {
-                self = .none
-            }
-        }
-    }
-    
-    private var neededLayoutType : NeededLayoutType = .requeryDataSourceCounts
-        
-    override func prepare()
-    {
-        super.prepare()
-
-        self.changesDuringCurrentUpdate = UpdateItems(with: [])
-        
-        self.neededLayoutType.update(with: {
-            switch self.neededLayoutType {
-            case .none: return true
-            case .updateHeaders: return self.performHeaderLayout()
-            case .relayout: return self.performRelayout()
-            case .requeryDataSourceCounts: return self.performRequeryDataSourceCountsAndLayout()
-            }
-        }())
-    }
-    
-    override func prepare(forCollectionViewUpdates updateItems: [UICollectionViewUpdateItem])
-    {
-        super.prepare(forCollectionViewUpdates: updateItems)
-        
-        self.changesDuringCurrentUpdate = UpdateItems(with: updateItems)
-    }
-    
-    //
-    // MARK: Finishing Layouts
-    //
-    
-    override func finalizeCollectionViewUpdates()
-    {
-        super.finalizeCollectionViewUpdates()
-        
-        self.changesDuringCurrentUpdate = UpdateItems(with: [])
-    }
-    
-    //
-    // MARK: Performing Layouts
-    //
-    
-    private func performHeaderLayout() -> Bool
-    {
-        return self.layoutResult.updateHeaders(in: self.collectionView!)
-    }
-    
-    private func performRelayout() -> Bool
-    {
-        return self.layoutResult.layout(
-            delegate: self.delegate,
-            in: self.collectionView!
-        )
-    }
-    
-    private func performRequeryDataSourceCountsAndLayout() -> Bool
-    {
-        self.previousLayoutResult = self.layoutResult
-        
-        self.layoutResult = LayoutInfo(
-            delegate: self.delegate,
-            appearance: self.appearance,
-            in: self.collectionView!
-        )
-        
-        return self.layoutResult.layout(
-            delegate: self.delegate,
-            in: self.collectionView!
-        )
-    }
-    
-    //
-    // MARK: UICollectionViewLayout Methods
-    //
-    
-    override var collectionViewContentSize : CGSize
-    {
-        return self.layoutResult.contentSize
-    }
-
-    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]?
-    {
-        return self.layoutResult.layoutAttributes(in: rect)
-    }
-
-    override func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes?
-    {
-        return self.layoutResult.layoutAttributes(at: indexPath)
-    }
-    
-    public override func layoutAttributesForSupplementaryView(ofKind elementKind: String, at indexPath: IndexPath) -> UICollectionViewLayoutAttributes?
-    {
-        return self.layoutResult.supplementaryLayoutAttributes(of: elementKind, at: indexPath)
-    }
-    
-    //
-    // MARK: UICollectionViewLayout Methods: Insertions & Removals
-    //
-
-    override func initialLayoutAttributesForAppearingItem(at itemIndexPath: IndexPath) -> UICollectionViewLayoutAttributes?
-    {
-        let wasInserted = self.changesDuringCurrentUpdate.insertedItems.contains(.init(newIndexPath: itemIndexPath))
-
-        if wasInserted {
-            let attributes = self.layoutResult.layoutAttributes(at: itemIndexPath)
-            
-            attributes.frame.origin.y -= attributes.frame.size.height
-            attributes.alpha = 0.0
-
-            return attributes
-        } else {
-            return super.initialLayoutAttributesForAppearingItem(at: itemIndexPath)
-        }
-    }
-
-    override func finalLayoutAttributesForDisappearingItem(at itemIndexPath: IndexPath) -> UICollectionViewLayoutAttributes?
-    {
-        let wasDeleted = self.changesDuringCurrentUpdate.deletedItems.contains(.init(oldIndexPath: itemIndexPath))
-
-        if wasDeleted {
-            let attributes = self.previousLayoutResult.layoutAttributes(at: itemIndexPath)
-
-            attributes.frame.origin.y -= attributes.frame.size.height
-            attributes.alpha = 0.0
-
-            return attributes
-        } else {
-            return super.finalLayoutAttributesForDisappearingItem(at: itemIndexPath)
-        }
-    }
-    
-    //
-    // MARK: UICollectionViewLayout Methods: Moving Items
-    //
-    
-    override func layoutAttributesForInteractivelyMovingItem(at indexPath: IndexPath, withTargetPosition position: CGPoint) -> UICollectionViewLayoutAttributes
-    {
-        let defaultAttributes = self.layoutResult.layoutAttributes(at: indexPath)
-        let attributes = super.layoutAttributesForInteractivelyMovingItem(at: indexPath, withTargetPosition: position)
-        
-        attributes.center.x = defaultAttributes.center.x
-        
-        return attributes
-    }
-}
+import Foundation
 
 
 extension ListViewLayout
@@ -463,7 +68,7 @@ internal extension ListViewLayout
         {
             self.collectionViewSize = .zero
             self.contentSize = .zero
-
+            
             self.appearance = Appearance()
             
             self.header = nil
@@ -476,13 +81,13 @@ internal extension ListViewLayout
             delegate : ListViewLayoutDelegate,
             appearance : Appearance,
             in collectionView : UICollectionView
-        )
+            )
         {
             let sectionCount = collectionView.numberOfSections
             
             self.collectionViewSize = collectionView.bounds.size
             self.contentSize = .zero
-
+            
             self.appearance = appearance
             
             self.header = {
@@ -504,7 +109,7 @@ internal extension ListViewLayout
                     layout: delegate.layoutForListFooter(in: collectionView)
                 )
             }()
-                        
+            
             self.sections = sectionCount.mapEach { sectionIndex in
                 
                 let itemCount = collectionView.numberOfItems(inSection: sectionIndex)
@@ -520,7 +125,7 @@ internal extension ListViewLayout
                             direction: appearance.direction,
                             layout: delegate.layoutForHeader(in: sectionIndex, in: collectionView)
                         )
-                    }(),
+                }(),
                     footer: {
                         guard delegate.hasFooter(in: sectionIndex, in: collectionView) else { return nil }
                         
@@ -529,7 +134,7 @@ internal extension ListViewLayout
                             direction: appearance.direction,
                             layout: delegate.layoutForFooter(in: sectionIndex, in: collectionView)
                         )
-                    }(),
+                }(),
                     columns: delegate.columnLayout(for: sectionIndex, in: collectionView),
                     items: itemCount.mapEach { itemIndex in
                         let indexPath = IndexPath(item: itemIndex, section: sectionIndex)
@@ -552,7 +157,7 @@ internal extension ListViewLayout
         func layoutAttributes(in rect: CGRect) -> [UICollectionViewLayoutAttributes]
         {
             var attributes = [UICollectionViewLayoutAttributes]()
-                        
+            
             if let header = self.header {
                 if rect.intersects(header.visibleFrame) {
                     attributes.append(header.layoutAttributes(with: SupplementaryKind.listHeader.indexPath(in: 0)))
@@ -564,7 +169,7 @@ internal extension ListViewLayout
                 guard rect.intersects(section.frame) else {
                     continue
                 }
-                                
+                
                 if let header = section.header {
                     if rect.intersects(header.visibleFrame) {
                         attributes.append(header.layoutAttributes(with: SupplementaryKind.sectionHeader.indexPath(in: sectionIndex)))
@@ -608,7 +213,7 @@ internal extension ListViewLayout
         func supplementaryLayoutAttributes(of kind : String, at indexPath : IndexPath) -> UICollectionViewLayoutAttributes?
         {
             let section = self.sections[indexPath.section]
-
+            
             switch SupplementaryKind(rawValue: kind)! {
             case .listHeader: return self.header?.layoutAttributes(with: indexPath)
             case .listFooter: return self.footer?.layoutAttributes(with: indexPath)
@@ -699,11 +304,11 @@ internal extension ListViewLayout
             
             return true
         }
-
+        
         func layout(
             delegate : ListViewLayoutDelegate,
             in collectionView : UICollectionView
-        ) -> Bool
+            ) -> Bool
         {
             guard collectionView.frame.size.isEmpty == false else {
                 return false
@@ -764,7 +369,7 @@ internal extension ListViewLayout
                 lastSectionMaxY += layout.padding.left
                 lastContentMaxY += layout.padding.left
             }
-
+            
             //
             // Sections
             //
@@ -894,7 +499,7 @@ internal extension ListViewLayout
                 
                 if isLast == false {
                     let additionalSectionSpacing = hasSectionFooter ? layout.interSectionSpacingWithFooter : layout.interSectionSpacingWithNoFooter
-
+                    
                     lastSectionMaxY += additionalSectionSpacing
                     lastContentMaxY += additionalSectionSpacing
                 }
@@ -921,7 +526,7 @@ internal extension ListViewLayout
                 lastContentMaxY = direction.maxY(for: footer.defaultFrame)
                 lastContentMaxY += layout.sectionHeaderBottomSpacing
             }
-                        
+            
             self.contentSize = direction.size(width: viewWidth, height: lastContentMaxY)
             
             self.updateHeaders(in: collectionView)
@@ -936,7 +541,11 @@ internal extension ListViewLayout
             }
         }
     }
-    
+}
+
+
+extension ListViewLayout.LayoutInfo
+{
     //
     // MARK: Layout Information
     //
@@ -961,7 +570,7 @@ internal extension ListViewLayout
             return CGRect(
                 origin: self.direction.point(x: self.x, y: self.y),
                 size: self.size
-             )
+            )
         }
         
         init(
@@ -971,7 +580,7 @@ internal extension ListViewLayout
             footer : SupplementaryItemLayoutInfo?,
             columns : Section.Columns,
             items : [ItemLayoutInfo]
-        )
+            )
         {
             self.direction = direction
             self.layout = layout
@@ -1049,7 +658,7 @@ internal extension ListViewLayout
     
     final class SupplementaryItemLayoutInfo
     {
-        let kind : SupplementaryKind
+        let kind : ListViewLayout.SupplementaryKind
         let direction : LayoutDirection
         let layout : HeaderFooterLayout
         
@@ -1062,7 +671,7 @@ internal extension ListViewLayout
             return CGRect(
                 origin: self.direction.point(x: self.x, y: self.y),
                 size: self.size
-             )
+            )
         }
         
         var visibleFrame : CGRect {
@@ -1072,7 +681,7 @@ internal extension ListViewLayout
             )
         }
         
-        init(kind : SupplementaryKind, direction : LayoutDirection, layout : HeaderFooterLayout)
+        init(kind : ListViewLayout.SupplementaryKind, direction : LayoutDirection, layout : HeaderFooterLayout)
         {
             self.kind = kind
             self.direction = direction
@@ -1116,8 +725,8 @@ internal extension ListViewLayout
             liveIndexPath : IndexPath,
             direction : LayoutDirection,
             layout : ItemLayout
-        )
-        {            
+            )
+        {
             self.delegateProvidedIndexPath = delegateProvidedIndexPath
             self.liveIndexPath = liveIndexPath
             
@@ -1176,82 +785,6 @@ fileprivate extension Array
     }
 }
 
-//
-// MARK: Update Information From Collection View Layout
-//
-
-fileprivate struct UpdateItems : Equatable
-{
-    let insertedSections : Set<InsertSection>
-    let deletedSections : Set<DeleteSection>
-    
-    let insertedItems : Set<InsertItem>
-    let deletedItems : Set<DeleteItem>
-    
-    init(with updateItems : [UICollectionViewUpdateItem])
-    {
-       var insertedSections = Set<InsertSection>()
-       var deletedSections = Set<DeleteSection>()
-
-       var insertedItems = Set<InsertItem>()
-       var deletedItems = Set<DeleteItem>()
-
-       for item in updateItems {
-            switch item.updateAction {
-            case .insert:
-                let indexPath = item.indexPathAfterUpdate!
-                
-                if indexPath.item == NSNotFound {
-                    insertedSections.insert(.init(newIndex: indexPath.section))
-                } else {
-                    insertedItems.insert(.init(newIndexPath: indexPath))
-                }
-                
-            case .delete:
-                let indexPath = item.indexPathBeforeUpdate!
-                
-                if indexPath.item == NSNotFound {
-                    deletedSections.insert(.init(oldIndex: indexPath.section))
-                } else {
-                    deletedItems.insert(.init(oldIndexPath: indexPath))
-                }
-                
-            case .move: break
-            case .reload: break
-            case .none: break
-                
-            @unknown default: fatalError()
-            }
-        }
-        
-        self.insertedSections = insertedSections
-        self.deletedSections = deletedSections
-        
-        self.insertedItems = insertedItems
-        self.deletedItems = deletedItems
-    }
-    
-    struct InsertSection : Hashable
-    {
-        var newIndex : Int
-    }
-    
-    struct DeleteSection : Hashable
-    {
-        var oldIndex : Int
-    }
-    
-    struct InsertItem : Hashable
-    {
-        var newIndexPath : IndexPath
-    }
-    
-    struct DeleteItem : Hashable
-    {
-        var oldIndexPath : IndexPath
-    }
-}
-
 
 fileprivate extension Int
 {
@@ -1266,4 +799,3 @@ fileprivate extension Int
         return mapped
     }
 }
-
